feat: move typeorm/modtree types out
