+++ packages/database/src/api/analyze.ts
@@ -1,7 +1,7 @@
 import { inspect } from 'util'
+import { AsyncFunction } from '@modtree/types'
 import { log } from '../cli'
 import { endpoint } from '../data-source'
-import { ModtreeFunction } from '../../types/modtree'
 import { db } from '../config'
 
 type Verbosity = 'none' | 'verbose' | 'normal'
@@ -91,11 +91,11 @@ function printResponse(response: any, verbosity: Verbosity) {
 /**
  * analyzes the output of an endpoint function
  *
- * @param {ModtreeFunction} callback
+ * @param {AsyncFunction} callback
  * @param {Verbosity} verbosity
  */
 export async function analyze<T>(
-  callback: ModtreeFunction<T>,
+  callback: AsyncFunction<T>,
   verbosity: Verbosity = 'normal'
 ) {
   const response = await endpoint(db, callback)
