+++ packages/types/index.d.ts
@@ -4,3 +4,4 @@ export * as NUSMods from './nusmods'
 export * as ModtreeApiResponse from './api-response'
 export * from './controller'
 export * from './repository'
+export * from './typeorm'
